Before changing Furniture Count Model

+++ ShopApi.DAL/Repositories/Address/AddressRepository.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
@@ -31,6 +32,8 @@ namespace ShopApi.DAL.Repositories.Address
 
         public async Task<bool> CreateAsync(Models.People.Address created)
         {
+            if (created == null)
+                return false;
             await _db.AddAsync(created);
             return true;
         }
@@ -38,7 +41,7 @@ namespace ShopApi.DAL.Repositories.Address
         public async Task<bool> UpdateAsync(int id, Models.People.Address updated)
         {
             var fromDb = await _db.AddressItems.FirstOrDefaultAsync(a => a.Id == id);
-            if (fromDb == null){return false;}
+            if (fromDb == null || updated == null){return false;}
 
             fromDb.City = updated.City;
             fromDb.House = updated.House;
@@ -52,6 +55,11 @@ namespace ShopApi.DAL.Repositories.Address
             var fromDb = await _db.AddressItems.FirstOrDefaultAsync(a => a.Id == id);
             if (fromDb == null){return false;}
 
+            if (await _db.PeopleItems.FirstOrDefaultAsync(p => p.Address.Id == id) != null)
+            {
+                throw new InvalidOperationException("Cannot remove address used by other entities in database. First remove binding");
+            }
+            
             _db.AddressItems.Remove(fromDb);
             return true;
         }
